imx6ul-telemetry: Configure SW3 voltage to 1.35V
Some imx6ul-telemetry are populated with a PF3000 model that
provides 1.2V in the SW3 by default.

SW3 output powers the DDR memory and the i.MX6UL datasheet specifies
that the minimum VDD DDR voltage is 1.283V.

Increase the SW3 voltage to the typical 1.35V to get into the valid
range.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Luan Rafael Carneiro <luan.rafael@ossystems.com.br>