omni700: Provide a MAC address
omni700 boards do not have the MAC address fuses programmed with
a valid MAC address.

This causes Linux to boot with a randomly generated MAC address.

Read the unique ID registers and run some logic on these registers
to provide a MAC address that is stored in U-Boot 'ethaddr' environment
variable.

With this approach, Linux always boots with the same MAC address that
is passed via U-Boot.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>