kontron: sl-mx6ul: Assign the eth1 MAC address
Currently, the eth1 MAC address is not assigned, which causes the kernel
to generate a random one for the second FEC port.

Fix this problem by assigning the eth1 MAC address directly by
reading the fuses.

The original implementation made by Kontron did not work
as it returned the MAC address of 00:00:00:00:00:00.

Signed-off-by: Fabio Estevam <festevam@gmail.com>